Feat: fix sentry
